sortable hintable

+++ ChecklistTracker/ViewModel/LayoutParams.cs
@@ -10,11 +10,11 @@ namespace ChecklistTracker.ViewModel
     internal class LayoutParams
     {
 
-        public double Width { get; set; }
-        public double Height { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
         public Thickness Padding { get; set; }
 
-        internal LayoutParams(double width, double height, Thickness padding)
+        internal LayoutParams(int width, int height, Thickness padding)
         {
             Width = width;
             Height = height;
